+++ src/test/java/domain/BoardTest.java
@@ -3,6 +3,9 @@ package domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import domain.board.Board;
+import domain.board.Line;
+import domain.path.PathValidator;
 import domain.piece.Pawn;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
