+++ src/test/java/domain/SquareTest.java
@@ -2,6 +2,7 @@ package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.board.Square;
 import domain.piece.EmptyPiece;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
